fix util tests add save/load tests
